Finished basic functionality for loading plugins that we developed ourselves.
