+++ FrontEnd/Modules/Templates/Controllers/DynamicContentController.cs
@@ -114,9 +114,9 @@ namespace FrontEnd.Modules.Templates.Controllers
         }
 
         /// <summary>
-        /// Get all possible components. These components should are retrieved from the assembly and should have the basetype CmsComponent&lt;CmsSettings, Enum&gt;
+        /// Get all possible components. These components should are retrieved from the assembly and any plugins and should have the base type CmsComponent&lt;CmsSettings, Enum&gt;
         /// </summary>
-        /// <returns>Dictionary of typeinfos and object attributes of all the components found in the GCL.</returns>
+        /// <returns>Dictionary of type infos and object attributes of all the components found in the GCL.</returns>
         private Dictionary<TypeInfo, CmsObjectAttribute> GetComponents()
         {
             var componentType = typeof(CmsComponent<CmsSettings, Enum>);
@@ -125,8 +125,7 @@ namespace FrontEnd.Modules.Templates.Controllers
             foreach (var assembly in loadedAssemblies)
             {
                 var typeInfoList = assembly.DefinedTypes.Where(
-                    type => type.BaseType != null
-                            && type.BaseType.IsGenericType
+                    type => type.BaseType is {IsGenericType: true}
                             && componentType.IsGenericType
                             && type.BaseType.GetGenericTypeDefinition() == componentType.GetGenericTypeDefinition()
                 ).OrderBy(type => type.Name).ToList();
@@ -137,22 +136,6 @@ namespace FrontEnd.Modules.Templates.Controllers
                 }
             }
 
-            /*var componentType = typeof(CmsComponent<CmsSettings, Enum>);
-            var assembly = componentType.Assembly;
-
-            var typeInfoList = assembly.DefinedTypes.Where(
-                type => type.BaseType != null
-                        && type.BaseType.IsGenericType
-                        && componentType.IsGenericType
-                        && type.BaseType.GetGenericTypeDefinition() == componentType.GetGenericTypeDefinition()
-            ).OrderBy(type => type.Name).ToList();
-
-            var resultDictionary = new Dictionary<TypeInfo, CmsObjectAttribute>();
-
-            foreach (var typeInfo in typeInfoList)
-            {
-                resultDictionary.Add(typeInfo, typeInfo.GetCustomAttribute<CmsObjectAttribute>());
-            }*/
             return resultDictionary;
         }
     }
