+++ FrontEnd/appsettings.Development.json
@@ -4,6 +4,7 @@
     "SecretsBaseDirectory": "Z:\\AppSettings\\WiserCore\\FrontEnd\\"
   },
   "FrontEnd": {
-    "ApiBaseUrl": "https://localhost:44349/"
+    "ApiBaseUrl": "https://localhost:44349/",
+    "PluginsDirectory": "C:\\Ontwikkeling\\Intern\\Wiser 3\\Api\\Plugins"
   }
 }
